独立しているstate
