+++ src/styles.css
@@ -1,7 +1,3 @@
-/* @tailwind base;
-@tailwind components;
-@tailwind utilities; */
-
 * {
   box-sizing: border-box;
 }
@@ -51,5 +47,29 @@ ul {
 }
 
 button {
-  margin-right: 10px;
+  display: block;
+  margin-bottom: 10px;
+}
+.Page > * {
+  float: left;
+  width: 50%;
+  padding: 10px;
+}
+h2 {
+  margin-top: 10px;
+  margin-bottom: 0;
+}
+h3 {
+  margin-top: 5px;
+  font-weight: normal;
+  font-size: 100%;
+}
+img {
+  width: 120px;
+  height: 120px;
+}
+button {
+  display: block;
+  margin-top: 10px;
+  margin-bottom: 10px;
 }
